project_main/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
google-generativeai>=0.8.0
google-genai>=1.0.0
langchain>=1.0.0
langchain-google-genai>=4.0.0
langgraph>=0.2.0
scikit-learn>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
jinja2>=3.1.0
python-multipart>=0.0.9
pyyaml>=6.0.0
